20201018(0)

+++ printf/src/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: yjung <yjung@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/16 18:20:05 by yjung             #+#    #+#             */
-/*   Updated: 2020/10/16 22:11:42 by yjung            ###   ########.fr       */
+/*   Updated: 2020/10/18 02:11:10 by yjung            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int		ft_printf(const char *format, ...)
 {
 	va_list	ap;
-	int		num;
+	size_t	num;
 
 	va_start(ap, format);
 	num = ft_print(0, format, ap);
